refactor(desktop): cleanup
