+++ harper-desktop/src/lib/client.ts
@@ -160,10 +160,6 @@ export class Client {
 		await invoke('set_integration_enabled', { bundleId, enabled });
 	}
 
-	static async getInstalledApplicationBundleIds(): Promise<string[]> {
-		return await invoke<string[]>('get_installed_application_bundle_ids');
-	}
-
 	static async getApplicationIconDataUrl(bundleId: string): Promise<string> {
 		return await invoke<string>('get_application_icon_data_url', { bundleId });
 	}
